refactor(KAN-424): move computeAccessTransition out of the admin route tree
F2 Defect 1 — eliminate the codebase's only src/lib/** -> src/app/**
import edge, and one of the two app->app cross-segment edges.

computeAccessTransition is the single source of truth for how a user moves
between access states, but it lived in src/app/admin/users/users-actions-shared.ts,
inside the admin UI route tree. src/lib/beta-access/flow.ts and
src/app/waitlist/actions.ts both reached into the admin console for it, which
made admin/access/beta-access/auth one inseparable knot.

It now lives in the neutral src/lib/access-model/ module — the seed of the
future `access` module (KAN-415).

This is motion, not an edit: the function body is byte-identical before and
after (verified by diff). The only signature change is the parameter type,
BulkAction -> AccessAction: the action vocabulary moves to access-model, while
the admin console's presentation of it (BULK_ACTIONS with its labels, danger
flags and menu order) stays behind, with BulkActionConfig.value typed as
AccessAction so the two cannot drift.

Also moved: the UserStatus / AccessTier axes, previously declared in
beta-access/flow.ts, which re-exports them so its public surface is unchanged.

Deliberately left in the admin console, per the ticket: BULK_ACTIONS, BULK_MAX,
EMAIL_CAP, UserFilter — those are genuinely admin's.

Verified:
  - grep for src/lib/** -> src/app/** import edges: 0 (was 1)
  - app->app cross-segment edges: waitlist->admin gone; [slug]->dashboard remains
  - unit suite 2564 passed / 217 suites, test count unchanged, no assertion touched
  - tsc --noEmit clean; eslint 0 errors
  - only a path update in tests/unit/access-model-transitions.test.ts

Scope note: KAN-424's other two defects (the organise WizardContact cycle and
the [slug] -> dashboard/profile edge) are NOT in this commit. Both are confined
to Luisa-owned UI paths (src/app/**/*.tsx and organise-fields.ts), so they need
her initiation and a UI-Change-Approved trailer. KAN-424 stays In Progress.
